+++ sync-for-expenses/RELEASES.md
@@ -393,3 +393,13 @@ Based on:
 - [go v4.0.0] sync-for-expenses
 ### Releases
 - [Go v4.0.0] https://github.com/codatio/client-sdk-go/releases/tag/sync-for-expenses/v4.0.0 - sync-for-expenses
+
+## 2024-02-04 18:09:42
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.166.0 (2.248.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v4.1.0] sync-for-expenses
+### Releases
+- [Go v4.1.0] https://github.com/codatio/client-sdk-go/releases/tag/sync-for-expenses/v4.1.0 - sync-for-expenses